Add auth-related repos and services
This is more of a rewrite of everything that's been in the
server-side code before.
There were also some changes all over the place. This is due to
having finally some time to apply the ruff linting rules and
formatting.
But the core of the commit stays the same:
there is now support for authorization, the auth logic
has been made to allow for heavy degree of DI.