Fix error on OTP verify
It fixes and error where otp code could not be sent to wrong access
to the OtpCode model.
The code has not been updated and worked on assumption that
OtpCode has an email attribute.
It could be prevented by a typechecker in pre-commit / CI.

Also, adds a default user to the database now on startup