Migrated to use PostgreSQL instead of an in-file database
The plan is to use pgvector for vector search - at least for a starter.
Big downside is that the server application won't work without providing
user's own configuration. It could be restored if need be.

Also, for making it easier to supply the credentials, the config model
has been extended to support providing settingss via env. variables.
It could be used for example when randomly generating username/password
for postgres during deployment.

To retain fully-operational status with no config there could be a MilvusLite client
provided, but it is not planned for now.