Add server.index module
Mainly: repositories and services for indexing have been added.
For now it works with postgresql+pgvector only.
The overall purpose is to allow for adding photos to the database.
ruff has been satisfied now.
Main change is a slight change in the directory layout. Embedder
repository has been under server.index module.
It's purpose is to store image embeddings along some metadata
in the vector database.