Add Embedder Registry
The purpose of the registry is to allow for defining various models
that can be used for the purpose of creating embeddings for the
similarity search.
The goal is to make it framework-agnostic and allow users to inject
models via configuration.
There are two supported ways:
1.via module (requires source code
modifications, used for built-in embedders),
2.via file path - using path to a python file. Useful
when one were to use it more as a library then fully owned
application.

There is also a default embedder sktr.py, which serves as an
example. The weights provided come from some training attempts
but on small dataset, having a very bad performer