Reshape dependency injection logic
Instead of writing the DI boilerplate for each singleton that has
been defined, everything's being done in the app startup.
This allows access to them in the lifespan, which is actually
important for setting up the `BackgroundEmbedder`.

The number of dependencies has been reduced. Now only these that
are used in the routes are defined.

Also, I replaced old observers with the background embedder