Add option to query images using IndexedImageRepository
This will be used for the purposes of the image gallery.

The IndexedImage model has been extended to split user visible
name and the file path. Distinguishing between these is reasonable
from security point of view, at least we could loosen the security
constraint in the future if we wanted to be "closer to fs".

Another important change is introduction of pytest fixtures
that allow for integration testing using a real postgreSQL database.