Create Docker configuration for development
Makes it a bit simpler to run the application to test things out.
Dockefiles have been added for the server and the frontend's
development server.

There were changes to the FastAPI server to allow for running it
in development mode via config.

Another important change is inclusion of gallery of sample images