Support thumbnails for images
Makes it so that thumbnails are fetched and displayed in the table.
There were some minor changes regarding thumbnails creation;
it has been made into a separate method of the service and now is
part of the "boostrapping" procedure.
Also fixed the bug, where query parameters were never passed into
the query; instead always the default query was being used.