Support nested directories with images
When recursively watching files, the gallery can be traversed
by directories. The images are also filtered by the directory.
Viewing gallery using all images is still supported, there is a
toggle for that.

There were also some changes w.r.t to path resolution on the
server side. Paths passed around were a bit convoluted.
The important thing is context (image/thumbnail) and relative
path (to watcher_root or thumbnail_root).

The system itself should not care about the exact paths,
the exact resolution is responsibility of the watcher.