Add admin panel
This is supposed to provide a UI for performing CRUD operations on
user, so that some new users can be added.

Now there is also a way to distinguish between roles of the user
currently using the app.
Access to the particular elements can be controlled in a
fine-grained manner using `RequireRole` helper component